test1

+++ Binomial.java
@@ -11,15 +11,15 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) { 
+	public static long binomial1(int n, int k) { 
 		if (k > n) return 0;
 		if (k == 0 || n == 0) return 1;
 		return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int [][] matrix = new int [n+1][k+1];
+	public static long binomial(int n, int k) {
+		long [][] matrix = new long [n+1][k+1];
 		for(int i = 0 ; i < matrix.length ; i++)
 			{
 				for(int j = 0 ; j < matrix[i].length ; j++)
@@ -31,7 +31,7 @@ public class Binomial {
 		return binomial(n, k, matrix);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
